Moving routes test to server testcase
